refactor: 소셜 서버에서 userinfo를 받는 dto 위치 변경
