+++ src/main/java/BE_Elixir/Elixir/domain/auth/service/OauthClient.java
@@ -1,6 +1,6 @@
 package BE_Elixir.Elixir.domain.auth.service;
 
-import BE_Elixir.Elixir.domain.auth.dto.SocialUserInfo;
+import BE_Elixir.Elixir.domain.auth.dto.social.SocialUserInfo;
 import BE_Elixir.Elixir.global.enums.LoginType;
 
 public interface OauthClient {
